+++ src/graph/mod.rs
@@ -3,7 +3,8 @@ pub mod dijkstra;
 pub mod dijkstra_restorepath;
 pub mod euler_tour;
 pub mod floyd_warshall;
-pub mod lowest_common_ancestor;
+pub mod lowest_common_ancestor_doubling;
+pub mod lowest_common_ancestor_rmq;
 pub mod minimun_spanning_tree_kruskal;
 pub mod minimun_spanning_tree_prim;
 pub mod util;
